Added a per-item SellPriceMultiplier which can override the DefaultSellPriceMultiplier within a shop config.
Added logic to scrape item sell prices from other shops which can be used if not StockPrice is specified. If no other shop price is found, it will still default to item.salePrice().

Fixed shop menu showing shared item quantity for all items.

Corrected erroneous (obsolete) Item comparison logic.

Improved check for CustomNPCExclusions.

Added config option to reschedule market day to the next day if it is cancelled for a festival, rain, or snow. It will not reschedule more than one day.

Reworked IsMarketday logic to include a user message on why it was cancelled, or if it has been rescheduled.

Re-enabled some less critical configuration options to be modifiable in-game.